SearchState MarkerID, AlleleID and SA index reduced to uint32

+++ libgramtools/include/common/utils.hpp
@@ -39,8 +39,8 @@ namespace gram {
     using Pattern = std::vector<Base>; /** A string of nucleotides is represented as a vector of `Base`s. */
     using Patterns = std::vector<Pattern>;
 
-    using Marker = uint64_t; /**< An integer >=5 describing a site or allele marker in the prg. */
-    using AlleleId = uint64_t; /**< An integer describing which allele is referred to within a given variant site. */
+    using Marker = uint32_t ; /**< An integer >=5 describing a site or allele marker in the prg. */
+    using AlleleId = uint32_t; /**< An integer describing which allele is referred to within a given variant site. */
 
     using VariantLocus = std::pair<Marker, AlleleId>; /**< A Variant site/`AlleleId` combination.*/
     using VariantSitePath = std::list<VariantLocus>; /**< A path through variant sites is a list of allele/site combinations. */
@@ -48,7 +48,7 @@ namespace gram {
 
     /** The suffix array (SA) holds the starting index of all (lexicographically sorted) cyclic permutations of the prg.
      * An `SA_Index` is an index into one such position.*/
-    using SA_Index = uint64_t;
+    using SA_Index = uint32_t;
     using SA_Interval = std::pair<SA_Index, SA_Index>; /**< A set of **contiguous** indices in the suffix array.*/
 
     /**
